chore:#12 remove test command; rename model files; rename schema files; move schemas and models to new folder
