+++ telegram_bot/bot/db/bullshit.model.js
@@ -1,4 +1,4 @@
 const mongoose = require('mongoose');
-const bullshitSchema = require('../schemas/Bullshit');
+const bullshitSchema = require('../../schemas/Bullshit');
 
 module.exports = new mongoose.model('Bullshit', bullshitSchema, 'bullshits');
