+++ telegram_bot/bot/db/user.model.js
@@ -1,4 +1,4 @@
 const mongoose = require('mongoose');
-const userSchema = require('../schemas/User');
+const userSchema = require('../../schemas/User');
 
 module.exports = new mongoose.model('User', userSchema, 'authorized_telegram_users');
